chore: use ts configs for federation testers
