+++ apps/tester-federation-v2/config.host-app.mts
@@ -1,5 +1,6 @@
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
+import type * as Repack from '@callstack/repack';
 
 import rspackConfig from './configs/rspack.host-app.mjs';
 import webpackConfig from './configs/webpack.host-app.mjs';
@@ -9,7 +10,7 @@ const __dirname = path.dirname(__filename);
 
 const useWebpack = Boolean(process.env.USE_WEBPACK);
 
-export default (env) => {
+export default (env: Repack.EnvOptions) => {
   env.context = __dirname;
 
   return useWebpack ? webpackConfig(env) : rspackConfig(env);
