+++ apps/tester-federation/configs/webpack.mini-app.mts
@@ -1,9 +1,7 @@
-// @ts-check
 import * as Repack from '@callstack/repack';
 import webpack from 'webpack';
 
-/** @type {(env: import('@callstack/repack').EnvOptions) => import('webpack').Configuration} */
-export default (env) => {
+export default Repack.defineWebpackConfig((env) => {
   const { mode, context, platform } = env;
 
   return {
@@ -15,7 +13,7 @@ export default (env) => {
     },
     output: {
       path: '[context]/build/mini-app/[platform]',
-      uniqueName: 'MF2Tester-MiniApp',
+      uniqueName: 'MFTester-MiniApp',
     },
     module: {
       rules: [
@@ -39,50 +37,53 @@ export default (env) => {
         ],
       }),
       // @ts-ignore
-      new Repack.plugins.ModuleFederationPluginV2({
+      new Repack.plugins.ModuleFederationPluginV1({
         name: 'MiniApp',
         filename: 'MiniApp.container.js.bundle',
         exposes: {
           './MiniAppNavigator': './src/mini/navigation/MainNavigator',
         },
-        dts: false,
         shared: {
           react: {
             singleton: true,
-            eager: false,
+            eager: true,
             requiredVersion: '19.0.0',
           },
           'react-native': {
             singleton: true,
-            eager: false,
+            eager: true,
             requiredVersion: '0.79.1',
           },
           '@react-navigation/native': {
             singleton: true,
-            eager: false,
+            eager: true,
             requiredVersion: '^6.1.18',
           },
           '@react-navigation/native-stack': {
             singleton: true,
-            eager: false,
+            eager: true,
             requiredVersion: '^6.10.1',
           },
           'react-native-safe-area-context': {
             singleton: true,
-            eager: false,
+            eager: true,
             requiredVersion: '^5.4.0',
           },
           'react-native-screens': {
             singleton: true,
-            eager: false,
+            eager: true,
             requiredVersion: '^4.10.0',
+          },
+          '@react-native-async-storage/async-storage': {
+            singleton: true,
+            eager: true,
+            requiredVersion: '^2.1.2',
           },
         },
       }),
-      // silence missing @react-native-masked-view optionally required by @react-navigation/elements
       new webpack.IgnorePlugin({
         resourceRegExp: /^@react-native-masked-view/,
       }),
     ],
   };
-};
+});
